클래스 점수 float로 변경

+++ models/class.js
@@ -16,7 +16,7 @@ module.exports = class Class extends Sequelize.Model {
 			},
 			// 강의 평가
 			class_score: {
-				type: Sequelize.INTEGER,
+				type: Sequelize.FLOAT,
 				allowNull: true,
 				defaultValue: 0,
 			},
